# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Meta Service",
    "description": "Wraps the Meta (Facebook) Graph API v22.0 for ad account management, performance reporting, and organic posting.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "serviceKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Service-to-service key (META_SERVICE_API_KEY)"
      }
    },
    "schemas": {
...
      "AuthorizeResponse": {
        "type": "object",
        "properties": {
          "authorizationUrl": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "authorizationUrl"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "nullable": true
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AdAccountResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "connectionId": {
            "type": "string",
            "format": "uuid"
          },
          "adAccountId": {
            "type": "string"
          },
          "accountName": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "nullable": true
          },
          "accountStatus": {
            "type": "number",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "connectionId",
          "adAccountId",
          "accountName",
          "currency",
          "timezone",
          "accountStatus",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "PatchAccountBody": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "isActive"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/auth/meta/authorize": {
      "get": {
        "summary": "Generate Meta OAuth login URL",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "required": true,
            "name": "redirectUri",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "label",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth authorization URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/accounts": {
      "get": {
        "summary": "List ad accounts across connections",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "boolean",
              "nullable": true,
              "default": true
            },
            "required": false,
            "name": "activeOnly",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "List of ad accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accounts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdAccountResponse"
                      }
                    }
                  },
                  "required": [
                    "accounts"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{adAccountId}": {
      "patch": {
        "summary": "Toggle ad account active/inactive",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "adAccountId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchAccountBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated ad account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdAccountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{adAccountId}/sync": {
      "post": {
        "summary": "Re-sync ad account metadata from Meta",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "adAccountId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Synced ad account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdAccountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}